requirements-airflow.txt
# requirements-airflow.txt

# 基礎 Airflow 版本
apache-airflow==2.8.4

# [修正] 顯式聲明所有需要的 Provider 包
apache-airflow-providers-celery
apache-airflow-providers-postgres
apache-airflow-providers-mysql
